check if a visit type has associated visits before deleting it

+++ app/Livewire/VisitTypes.php
@@ -62,7 +62,10 @@ class VisitTypes extends Component
 
     public function delete(VisitType $visitType)
     {
-        // Todo: Check if the visit type is associated with any visit
+        if ($visitType->visits->count()) {
+            return;
+        }
+
         $visitType->delete();
     }
